Execute cgit in the request method instead of needing another server
This patch, which seems to work even if without unit tests so far,
aims to cut a good deal of code and complexity from the application.

First: delegate to Catalyst and HTTP::Message the encoding/decoding.
This should be fine, as far as I could see. I mean, they're supposed
to be smarted than me anyway.

Second: Instead of the need of a server just to run cgi, and having to
deal with ports, etc., set the environment variables needed by CGI and
execute it, capture the output and parse it. Then relay.

TODO: remove code which is not needed anymore

Address: #245